controller script to send CAN packets
This will send the CAN packets using the commands like can_send instead of using a terminal. This command is binded to the custom GUI key controls. pressing up key send throttle messages, left key sends steer left turn and so on.. This needs to be made more efficient by deciding the proper no of messages to be sent on one press of custom GUI key.